+++ sources/bonus/init/struct/init.c
@@ -6,7 +6,7 @@
 /*   By: twang <twang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/04 16:48:31 by twang             #+#    #+#             */
-/*   Updated: 2023/08/24 14:21:47 by twang            ###   ########.fr       */
+/*   Updated: 2023/08/24 14:51:50 by twang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void	set_vector(t_vector *vector, int x, int y)
 	vector->y = y;
 }
 
-void	set_vector_f(t_vector_f *vector, int x, int y)
+void	set_vector_f(t_vector_f *vector, float x, float y)
 {
 	vector->x = x;
 	vector->y = y;
